fix: add ratios to default_column function and move IntensityNameType to constants
