+++ backend/protzilla/constants/intensity_types.py
@@ -1,0 +1,23 @@
+from enum import Enum
+
+
+class IntensityType(Enum):
+    IBAQ = "iBAQ"
+    INTENSITY = "Intensity"
+    LFQ_INTENSITY = "LFQ intensity"
+    RATIO_HL = "Ratio H/L"
+    RATIO_LH = "Ratio L/H"
+    RATIO_HL_NORMALIZED = "Ratio H/L normalized"
+    RATIO_LH_NORMALIZED = "Ratio L/H normalized"
+
+
+class IntensityNameType(Enum):
+    INTENSITY = "Intensity"
+    MAXLFQ_TOTAL_INTENSITY = "MaxLFQ Total Intensity"
+    MAXLFQ_INTENSITY = "MaxLFQ Intensity"
+    TOTAL_INTENSITY = "Total Intensity"
+    MAXLFQ_UNIQUE_INTENSITY = "MaxLFQ Unique Intensity"
+    UNIQUE_SPECTRAL_COUNT = "Unique Spectral Count"
+    UNIQUE_INTENSITY = "Unique Intensity"
+    SPECTRAL_COUNT = "Spectral Count"
+    TOTAL_SPECTRAL_COUNT = "Total Spectral Count"
