+++ backend/protzilla/importing/example_dataset_import.py
@@ -4,7 +4,7 @@ from protzilla.constants.paths import (
     EXAMPLE_DATASET_EVIDENCE_FILE,
 )
 from protzilla.importing.import_utils import FeatureOrientationType
-from protzilla.constants.intensity_type import IntensityType
+from protzilla.constants.intensity_types import IntensityType
 from protzilla.importing.metadata_import import metadata_import_method
 from protzilla.importing.ms_data_import import max_quant_import
 from protzilla.importing.peptide_import import evidence_import
